missed some files

+++ src/main/java/com/thefitnation/service/UserDemographicService.java
@@ -12,11 +12,14 @@ import com.thefitnation.web.rest.util.HeaderUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Objects;
 import java.util.Optional;
 
 import static org.hibernate.id.IdentifierGenerator.ENTITY_NAME;
@@ -61,9 +64,12 @@ public class UserDemographicService {
 
         if (AccountAuthTool.isAdmin(user)) {
             UserDemographic dbUserDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(userDemographic.getUser().getId());
-            userDemographic.setId(dbUserDemographic.getId());
+            if (dbUserDemographic == null) {
+                userDemographic.setId(null);
+            }
             userDemographic = userDemographicRepository.save(userDemographic);
         } else {
+            userDemographic.setUser(user);
             UserDemographic dbUserDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(user.getId());
             if (dbUserDemographic == null) {
                 userDemographic.setId(null);
@@ -86,7 +92,13 @@ public class UserDemographicService {
     @Transactional(readOnly = true)
     public Page<UserDemographicDTO> findAll(Pageable pageable) {
         log.debug("Request to get all UserDemographics");
-        Page<UserDemographic> result = userDemographicRepository.findAll(pageable);
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        Page<UserDemographic> result;
+        if (AccountAuthTool.isAdmin(user)) {
+            result = userDemographicRepository.findAll(pageable);
+        } else {
+            return null;
+        }
         return result.map(userDemographicMapper::userDemographicToUserDemographicDTO);
     }
 
@@ -99,21 +111,26 @@ public class UserDemographicService {
     @Transactional(readOnly = true)
     public UserDemographicDTO findOne(Long id) {
         log.debug("Request to get UserDemographic : {}", id);
-        UserDemographic userDemographic = userDemographicRepository.findOneWithEagerRelationships(id);
-        UserDemographicDTO userDemographicDTO = userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
-        return userDemographicDTO;
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        UserDemographic userDemographic;
+        if (AccountAuthTool.isAdmin(user)) {
+            userDemographic = userDemographicRepository.findOneWithEagerRelationships(id);
+        } else {
+            userDemographic = userDemographicRepository.findOneWithEagerRelationships(id);
+            if (!Objects.equals(userDemographic.getUser().getId(), user.getId())) {
+                return null;
+            }
+        }
+        return userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
     }
 
     @Transactional(readOnly = true)
     public UserDemographicDTO findOneByUser() {
-        log.debug("Request to get UserDemographic by User : {}");
-        Optional<User> user = userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin());
-        if (!user.isPresent()) {
-            return null;
-        }
-        UserDemographic userDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(user.get().getId());
-        UserDemographicDTO userDemographicDTO = userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
-        return userDemographicDTO;
+        log.debug("Request to get UserDemographic : {}");
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        UserDemographic userDemographic;
+        userDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(user.getId());
+        return userDemographicMapper.userDemographicToUserDemographicDTO(userDemographic);
     }
 
     /**
@@ -123,6 +140,9 @@ public class UserDemographicService {
      */
     public void delete(Long id) {
         log.debug("Request to delete UserDemographic : {}", id);
-        userDemographicRepository.delete(id);
+        User user = AccountAuthTool.getLoggedInUser(userRepository);
+        if (AccountAuthTool.isAdmin(user)) {
+            userDemographicRepository.delete(id);
+        }
     }
 }
